UWIPMC_INITIALIZE_STATIC_SEMAPHORE()
Create a common macro in IPMC.h to handle boot-time initialization of
FreeRTOS semaphores that are statically declared.  FreeRTOS seems to
lack a static initializer macro that can simply be assigned to these
types, requiring intialization with runtime code instead.

This macro will create a function, automatically run before main(), that
calls the appropriate FreeRTOS initializer.

The priority of this initializer relative to other similar functions may
change in the future, as other needs appear.