Add init functions.  Connect printf/xil_printf to uart_ps0.write().
Add init functions to IPMC.cc, which are responsible for driver
initialization and object construction.

Arrange for proper initialization of the uart_ps0 driver.

Hijack xil_printf by #define, and alter printf/vprintf to direct output
to uart_ps0.write().  xil_printf() is now DEPRECATED.  Do not use it.
It will likely be changed to a configASSERT(0) in the future.