Remove mutexes from WaitList and Event threading primitives.
This commit removes the mutexes from the WaitList and Event threading
primitives, to allow them to be used in more circumstances, such as when
the scheduler is suspended.

This commit also fixes a race condition on WaitList.wake() with
wait()ing subscriptions timing out while the wake is in progress, and
unifies the wakeFromISR and wake functions.