Add LogTree and TraceBuffer facilities.
This will allow customizable logging using a log facility hierarchy.
The verbosity of output to various channels may be tuned individually
for each subtree in the hierarchy.

The tracebuffer facility will log messages at extreme verbosity to a
large dedicated memory buffer which may be dumped over JTAG or via other
means and parsed using the supplied python script.  Although logging
from ISRs is not permissible, it is possible to record messages to the
tracebuffer from within ISRs, to allow for the recording of debug
information from ISRs.