Improve version string embedding technique.
The old version string embedding relied on defines referenced in each
individual file, however this did not account for make dependency
optimization, and version information would be updated only when make
clean was run or when the file using the defines was modified.

This new scheme for version string embedding creates a separate
version.cc object containing const char* strings containing the version
information, which are exported and referenced by the linker.  This
object is not compiled as version.o, but rather
version.${md5_of_git_info}.o, ensuring it is never out of date, while
not triggering re-linking of the project if no changes were made.