Perform SPI_EEPROM reads with a single large malloced buffer.
Perform SPI_EEPROM reads with a single large buffer allocated via
pvPortMalloc, rather than using smaller stack allocated buffers to read
a page at a time.  This has cut the duration of a 256k read by 79%.
This has the cost of the slight potential for heap fragmentation.  This
potential, however we have sufficient memory available to us for it to
be relatively self-healing even if it occurs.