The initialization tasks MUST run at TASK_PRIORITY_WATCHDOG
Without this, it is possible for other processes within the IPMC to
start moving before all setup is completed, which may have unpredictable
results.  If an external library such as lwip insists on blocking or
parallel operation during startup, it should be dispatched to another
thread for initialization after other core initialization is complete.
Things can be further chained off of this by way of a callback parameter
to this thread.