Increase initial stack size from 8k to 128k.
The stack was being overflowed due to the C++ std::map initalizers
copying a very large set of initialization_vectors found in IPMC.cc onto
the stack as a part of the initialization process.  This occured before
main(), during the static constructors portion of the startup sequence,
on the only stack not guarded by the FreeRTOS scheduler.  Hopefully 128k
will solve this problem for the future.  We can certainly afford it.