Fix version.cc build dependency checks and add build host.
Since the version.cc formerly contained only git related information,
the Makefile was configured to only rebuild it when that information
changed.  Now that the build date was added in 6592fde, this is no
longer correct, and if working in a dirty repository, the git
information will not change between builds, and the build date will not
be updated, likely leading to confusion.  This change corrects this
behavior.  This optimization has been disabled and the version.cc and
therefore main elf file will now ALWAYS be rebuilt, regardless of the
presence or absence of changes.