IPMIFormats: Fix offset handling in the type/length string parser broken in 1b5a3efe
When this was converted from a copy to a const reference, I changed the
Packed 6-bit ASCII's starting offset to 1, but this results in incorrect
parsing becuase it offsets the state machine by one.  I also failed to
correctly update the Raw ASCII or Unicode handler.  Both of these
issues are now corrected.