We cannot allow abort() to return, even if configASSERT(0) is stepped out of.
abort() is defined with attribute noreturn, so GCC will not generate a
function epilogue for it.  configASSERT(0) can be stepped out of within
the debugger.  abort() literally can't.  We will therefore `while(1);`
to satisfy the warning, and ensure later programmers do not accidentally
attempt to do things that are insane on an assembly level.