Add strong image validation support, and primary/backup partition scheme.
Images are now validated by reading the ZYNQ part out of the raw
bitstream, as well as the ZYNQ device ID register.

Images are now validatd by confirming that the git tag in the embedded
version information matches the expected tag, if configured.

Images are now validated when switching boot target as well as when
uploading.

Flash partitions are now fallback, primary/backup, backup/primary, test.
Updates automatically program the next image slot, and boottarget can be
used to undo an update.